pyproject.toml
[tool.poetry]
name = "mindai"
version = "0.1.0"
description = "An agentic AI with Active Memory"
authors = ["Martin Bukowski <4502632+54rt1n@users.noreply.github.co>"]
license = "CC-BY-NC-SA-4.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
chardet = "^5.2.0"
bullmq = "2.9.4"
redis = "5.1.1"
semver = "2.13.0"
click = "*"
nltk = "*"
numpy = "<2"
openai = "^1.58.1"
pandas = "*"
python-dotenv = "*"
scikit-learn = "*"
torch = "*"
transformers = "*"
sentence-transformers = "^3.2.0"
watchdog = "^5.0.3"
wonderwords = "2.2.0"
fastapi = "0.115.0"
sse-starlette = "2.1.3"
starlette = "0.38.6"
uvicorn = "0.31.1"
trafilatura = "1.12.2"
lxml_html_clean = "0.3.1"
python-multipart = "0.0.17"
tantivy = "^0.22.0"
faiss-cpu = "^1.9.0.post1"
pydantic = "^2.10.6"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
